requirements.txt
langchain
langchain_openai
langgraph
python-dotenv
tavily-python
numpy
faiss-cpu
# pickle5